# backend/openapi.yaml
openapi: 3.1.0
info:
  title: AI Threat Modeler API
  description: |
    API for the AI Threat Modeler Dashboard - A comprehensive application security platform 
    with threat modeling and AI-powered chat capabilities.
    
    ## Authentication
    Most endpoints require JWT Bearer token authentication. Obtain a token by logging in 
    via the `/api/auth/login` endpoint.
    
    ## User Roles
    - **Admin**: Full access to all features including user management and settings
    - **Operator**: Can schedule and manage their own jobs (threat modeling)
    - **Auditor**: Read-only access to view all jobs and reports
    
  version: 0.9.0
  contact:
    name: Sam Li
  license:
    name: MIT
servers:
  - url: http://localhost:3001
    description: Local development server
  - url: http://localhost:3001/api
    description: Local development API
...
security:
  - BearerAuth: []
paths:
...
  /api/chat:
    post:
      tags:
        - Chat
      summary: Send chat message
      description: |
        Send a message to the AI chat agent and receive a response.
        Sessions are maintained per user. Send "/end" to terminate the session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
                  description: User message
                  example: What are common SQL injection vulnerabilities?
                role:
                  type: string
                  description: AI agent role
                  default: simple_query_agent
                  example: simple_query_agent
      responses:
        '200':
          description: Chat response received
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  response:
                    type: string
                    description: AI response
                  role:
                    type: string
                  sessionActive:
                    type: boolean
                  sessionEnded:
                    type: boolean
                    description: Set to true if "/end" was sent
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/chat/end:
    post:
      tags:
        - Chat
      summary: End chat session
      description: Explicitly terminate the current chat session
      responses:
        '200':
          description: Session ended
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                  sessionEnded:
                    type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/chat/session:
    get:
      tags:
        - Chat
      summary: Get session status
      description: Check if user has an active chat session
      responses:
        '200':
          description: Session status retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  hasSession:
                    type: boolean
                  sessionActive:
                    type: boolean
                  lastActivity:
                    type: string
                    format: date-time
                    nullable: true
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Error message
        message:
          type: string
          example: Detailed error description
        code:
          type: string
          description: Optional error code
          example: USER_NOT_FOUND
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Bad request
            message: Invalid parameters provided
    Unauthorized:
      description: Unauthorized - missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Access token required
            message: Please provide a valid authentication token
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Internal server error
            message: An unexpected error occurred